# libs/shared/apis/v1/synaptiq-apis/src/main/java/com/synaptiq/infrastructure/in/web/api/SchemaRegistryApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.synaptiq.infrastructure.in.web.api;

import com.synaptiq.infrastructure.in.web.dto.CollectionListResponse;
import java.util.Map;
import org.springframework.lang.Nullable;
import com.synaptiq.infrastructure.in.web.dto.QueryResultResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-05T21:15:25.464614100-05:00[America/Chicago]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "SchemaRegistry", description = "Schema Registry — collection inspection, dynamic queries")
public interface SchemaRegistryApi {

    String PATH_INFER_SCHEMA = "/api/v1/schema-registry/infer/{collection}";
    /**
     * GET /api/v1/schema-registry/infer/{collection} : Infer collection schema
     *
     * @param collection  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Inferred schema (status code 200)
     */
    @Operation(
        operationId = "inferSchema",
        summary = "Infer collection schema",
        tags = { "SchemaRegistry" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Inferred schema", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Map.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SchemaRegistryApi.PATH_INFER_SCHEMA,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<Map<String, String>>> inferSchema(
        @Parameter(name = "collection", description = "", required = true, in = ParameterIn.PATH) @PathVariable("collection") String collection,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_LIST_COLLECTIONS = "/api/v1/schema-registry/collections";
    /**
     * GET /api/v1/schema-registry/collections : List MongoDB collections
     *
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @return Collections (status code 200)
     */
    @Operation(
        operationId = "listCollections",
        summary = "List MongoDB collections",
        tags = { "SchemaRegistry" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Collections", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CollectionListResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = SchemaRegistryApi.PATH_LIST_COLLECTIONS,
        produces = { "application/json" }
    )
    Mono<ResponseEntity<CollectionListResponse>> listCollections(
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );


    String PATH_QUERY_COLLECTION = "/api/v1/schema-registry/query/{collection}";
    /**
     * POST /api/v1/schema-registry/query/{collection} : Query a collection
     *
     * @param collection  (required)
     * @param xTenantID The tenant identifier for multi-tenant isolation (optional)
     * @param limit  (optional, default to 20)
     * @param requestBody  (optional)
     * @return Query results (status code 200)
     */
    @Operation(
        operationId = "queryCollection",
        summary = "Query a collection",
        tags = { "SchemaRegistry" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Query results", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = QueryResultResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = SchemaRegistryApi.PATH_QUERY_COLLECTION,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    Mono<ResponseEntity<QueryResultResponse>> queryCollection(
        @Parameter(name = "collection", description = "", required = true, in = ParameterIn.PATH) @PathVariable("collection") String collection,
        @Size(min = 1, max = 100) @Parameter(name = "X-Tenant-ID", description = "The tenant identifier for multi-tenant isolation", in = ParameterIn.HEADER) @RequestHeader(value = "X-Tenant-ID", required = false) @Nullable String xTenantID,
        @Parameter(name = "limit", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false, defaultValue = "20") Integer limit,
        @Parameter(name = "request_body", description = "") @Valid @RequestBody(required = false) Mono<Map<String, Object>> requestBody,
        @Parameter(hidden = true) final ServerWebExchange exchange
    );

}