# libs/shared/openapi-spec/src/main/resources/openapi/bundled-synaptiq-v1.yaml
openapi: 3.1.0
info:
  title: Synaptiq API
  description: |
    AI-Powered Conversational Commerce Platform — REST API

    Synaptiq enables businesses to deploy intelligent, context-aware AI shopping assistants
    with full catalog management, conversation orchestration, workflow automation,
    and multi-tenant governance. This specification covers all bounded contexts:
    Authentication, Tenants, Catalog, Chat, Actions, Workflows, Analytics,
    Branding, Tenant Configuration, and Schema Registry.
  version: 1.0.0
  contact:
    name: Synaptiq Team
    url: https://synaptiq.dev
  license:
    name: Proprietary
    url: https://synaptiq.dev/license
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.synaptiq.dev
    description: Production server
...
paths:
...
  /api/v1/schema-registry/collections:
    get:
      operationId: listCollections
      summary: List MongoDB collections
      tags:
        - SchemaRegistry
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Collections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
  /api/v1/schema-registry/infer/{collection}:
    get:
      operationId: inferSchema
      summary: Infer collection schema
      tags:
        - SchemaRegistry
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: collection
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Inferred schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InferredSchemaResponse'
  /api/v1/schema-registry/query/{collection}:
    post:
      operationId: queryCollection
      summary: Query a collection
      tags:
        - SchemaRegistry
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: collection
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Query results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResultResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: The tenant identifier for multi-tenant isolation
      schema:
        type: string
        minLength: 1
        maxLength: 100
  schemas:
...
    CollectionListResponse:
      type: object
      properties:
        collections:
          type: array
          items:
            type: string
    InferredSchemaResponse:
      type: object
      additionalProperties:
        type: string
        description: Field name to inferred type mapping
    QueryResultResponse:
      type: object
      properties:
        documents:
          type: array
          items:
            type: object
            additionalProperties: true
        total:
          type: integer
...